==> 5_plot_brain_figures.ipynb <==
import pandas as pd
import numpy as np
from nilearn import plotting
import matplotlib.pyplot as plt
from nilearn import surface
from nilearn import datasets

def plot_DK_brain_figures(diff_path,colname,mytype,tmax=None,sig_t=None):
    
    lh_atlas = surface.load_surf_data('raw_data/atlas/lh.aparc.annot')
    rh_atlas = surface.load_surf_data('raw_data/atlas/rh.aparc.annot')

    imaging_diff = pd.read_csv(diff_path)
    regions = imaging_diff['regions'].values.tolist()
    mapping_matrix = pd.read_csv('raw_data/atlas/value_to_region_table.csv').set_index('regions')
    mapping_matrix = mapping_matrix.loc[regions]
    region_ids = mapping_matrix['value'].values
    plot_values = imaging_diff[colname].values

    s1 = lh_atlas.shape
    l_mask = np.zeros(s1)
    s2 = rh_atlas.shape
    r_mask = np.zeros(s2)
    
    for i in range(len(regions)):
        region = regions[i]
        region_id = region_ids[i]
        if region[4:6] == 'lh':
            l_mask[lh_atlas == region_id] = plot_values[i]
        else:
            r_mask[rh_atlas == region_id] = plot_values[i]
        
    fsaverage = datasets.fetch_surf_fsaverage(mesh='fsaverage5')
    
    plotting.plot_surf_stat_map(fsaverage['pial_left'], stat_map=l_mask,
                                hemi='left', view='lateral', cmap='RdBu', colorbar=True,vmax=tmax,
                                bg_map=fsaverage['sulc_left'], bg_on_data=True,
                                darkness=.5,output_file='results/brain_figures/'+mytype+'_T_map_left_lateral.png')

    plotting.plot_surf_stat_map(fsaverage['pial_left'], stat_map=l_mask,
                                hemi='left', view='medial', cmap='RdBu', colorbar=True,vmax=tmax,
                                bg_map=fsaverage['sulc_left'], bg_on_data=True,
                                darkness=.5,output_file='results/brain_figures/'+mytype+'_T_map_left_medial.png')


    plotting.plot_surf_stat_map(fsaverage['pial_right'], stat_map=r_mask,
                                hemi='right', view='medial', cmap='RdBu', colorbar=True,vmax=tmax,
                                bg_map=fsaverage['sulc_right'], bg_on_data=True,
                                darkness=0.5,output_file='results/brain_figures/'+mytype+'_T_map_right_medial.png')

    plotting.plot_surf_stat_map(fsaverage['pial_right'], stat_map=r_mask,
                                hemi='right', view='lateral', cmap='RdBu', colorbar=True,vmax=tmax,
                                bg_map=fsaverage['sulc_right'], bg_on_data=True,
                                darkness=.5,output_file='results/brain_figures/'+mytype+'_T_map_right_lateral.png')

    if sig_t != None:
        plotting.plot_surf_stat_map(fsaverage['pial_left'], stat_map=l_mask,
                                hemi='left', view='lateral', cmap='RdBu', colorbar=True,
                                bg_map=fsaverage['sulc_left'], bg_on_data=True, threshold=sig_t,vmax=tmax,
                                darkness=.5,output_file='results/brain_figures/sig_'+mytype+'_T_map_left_lateral.png')

        plotting.plot_surf_stat_map(fsaverage['pial_left'], stat_map=l_mask,
                                    hemi='left', view='medial', cmap='RdBu', colorbar=True,
                                    bg_map=fsaverage['sulc_left'], bg_on_data=True,threshold=sig_t,vmax=tmax,
                                    darkness=.5,output_file='results/brain_figures/sig_'+mytype+'_T_map_left_medial.png')


        plotting.plot_surf_stat_map(fsaverage['pial_right'], stat_map=r_mask,
                                    hemi='right', view='medial', cmap='RdBu', colorbar=True,
                                    bg_map=fsaverage['sulc_right'], bg_on_data=True,threshold=sig_t,vmax=tmax,
                                    darkness=.5,output_file='results/brain_figures/sig_'+mytype+'_T_map_right_medial.png')

        plotting.plot_surf_stat_map(fsaverage['pial_right'], stat_map=r_mask,
                                    hemi='right', view='lateral', cmap='RdBu', colorbar=True,
                                    bg_map=fsaverage['sulc_right'], bg_on_data=True,threshold=sig_t,vmax=tmax,
                                    darkness=.5,output_file='results/brain_figures/sig_'+mytype+'_T_map_right_lateral.png')

CT_corr_path = 'results/imaging_associations/CT_corr_PC1.csv'
CSA_corr_path = 'results/imaging_associations/CSA_corr_PC1.csv'
ALFF_corr_path = 'results/imaging_associations/ALFF_corr_PC1.csv'
ReHo_corr_path = 'results/imaging_associations/ReHo_corr_PC1.csv'

plot_DK_brain_figures(CT_corr_path,'r','CT',tmax=0.05,sig_t=0.02)
plot_DK_brain_figures(CSA_corr_path,'r','CSA',tmax=0.05,sig_t=0.02)
plot_DK_brain_figures(ALFF_corr_path,'r','ALFF',tmax=0.05,sig_t=0.016)
plot_DK_brain_figures(ReHo_corr_path,'r','ReHo',tmax=0.05,sig_t=0.016)

funcPC_path = 'results/imaging_associations/func_PCs.csv'
plot_DK_brain_figures(funcPC_path,'PC1','funcPC1',tmax=2)

# GWASPC_path = 'results/expression_associations/genes_PCs.csv'
# plot_DK_brain_figures(GWASPC_path,'PC1','GWASPC1',tmax=2.8)
# plot_DK_brain_figures(GWASPC_path,'PC2','GWASPC2',tmax=2.8)

allgenePC_path = 'results/expression_associations/gene_PCs.csv'
plot_DK_brain_figures(allgenePC_path,'PC1','allgenesPC1',tmax=2.6,sig_t=1.96)
plot_DK_brain_figures(allgenePC_path,'PC2','allgenesPC2',tmax=2.6,sig_t=1.96)

# CT_pos_PC_path = 'results/expression_associations/CT_pos_genes_PCs.csv'
# plot_DK_brain_figures(CT_pos_PC_path,'PC1','CT_pos_PC1',tmax=3)

# CT_neg_PC_path = 'results/expression_associations/CT_neg_genes_PCs.csv'
# plot_DK_brain_figures(CT_neg_PC_path,'PC1','CT_neg_PC1',tmax=3)

# funcPC1_pos_PC_path = 'results/expression_associations/funcPC1_pos_genes_PCs.csv'
# plot_DK_brain_figures(funcPC1_pos_PC_path,'PC1','funcPC1_pos_PC1',tmax=3)

# funcPC1_neg_PC_path = 'results/expression_associations/funcPC1_neg_genes_PCs.csv'
# plot_DK_brain_figures(funcPC1_neg_PC_path,'PC1','funcPC1_neg_PC1',tmax=3)

CADM2_path = 'results/expression_associations/CADM2_exp.csv'
plot_DK_brain_figures(CADM2_path,'CADM2','CADM2',tmax=3.2)
